rpc: move web3.js extensions to internal/web3ext
